DTOs e mappers

+++ src/main/java/com/contas107/service/LancamentoService.java
@@ -2,8 +2,8 @@ package com.contas107.service;
 
 import java.util.Date;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.Calendar;
+import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.ZoneId;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,8 +12,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.server.ResponseStatusException;
 
+import com.contas107.DTO.LancamentoRequestDTO;
+import com.contas107.DTO.LancamentoResponseDTO;
+import com.contas107.mapstruct.LancamentoMapper;
+import com.contas107.model.Banco;
 import com.contas107.model.Lancamento;
+import com.contas107.repository.BancoRepository;
 import com.contas107.repository.LancamentoRepository;
+import com.mysql.cj.util.StringUtils;
 
 @Service
 public class LancamentoService {
@@ -21,56 +27,89 @@ public class LancamentoService {
     @Autowired
     private LancamentoRepository lancamentoRepository;
 
-    public List<Lancamento> obterTodosLancamentos(){
-        List <Lancamento> retorno = lancamentoRepository.findAll();
+    @Autowired
+    private BancoRepository bancoRepository;
+
+    @Autowired
+    private LancamentoMapper lancamentoMapper;
+
+    public List<LancamentoResponseDTO> obterTodosLancamentos(){
+        List <LancamentoResponseDTO> retorno = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.findAll());
         if (retorno.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado. 🙁");
         }
         return retorno;
     }
 
-    public Lancamento obterLancamentoPorId(Long id) {
-        return lancamentoRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Lançamento com o ID: " + id + " não encontrado. ⛔"));
+    public LancamentoResponseDTO obterLancamentoPorId(Long id) {
+        LancamentoResponseDTO retorno = lancamentoMapper.paraLancamentoResponseDTO(lancamentoRepository.findById(id).orElse(null));
+        if (retorno == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Lançamento com o ID: " + id + " não encontrado.");
+        }
+        return retorno;
     }   
 
-    public Lancamento salvarLancamento(@RequestBody Lancamento lancamento) {
-        try {
-            lancamento.setDataLancamento(new Date().toInstant().atZone(ZoneId.systemDefault()).toLocalDate());
-            return lancamentoRepository.save(lancamento);
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Erro ao salvar o Lançamento. 🙁", e);
+    public LancamentoResponseDTO salvarLancamento(@RequestBody LancamentoRequestDTO lancamento) {
+        if (lancamento.getBanco() == null 
+            || lancamento.getBanco().getId() == null 
+            || bancoRepository.findById(lancamento.getBanco().getId()).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Banco não encontrado. ⛔");
         }
+
+        if (StringUtils.isNullOrEmpty(lancamento.getDataLancamento().toString())) {
+            lancamento.setDataLancamento(LocalDate.now(ZoneId.of("UTC"))); 
+        }
+
+        LancamentoResponseDTO retorno = lancamentoMapper.paraLancamentoResponseDTO(lancamentoRepository.save(lancamentoMapper.paraLancamentoEntity(lancamento)));
+        if (retorno == null) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Erro ao salvar o Lançamento. 🙁");
+        }
+        return retorno;
+
     }
 
-    public Lancamento atualizarLancamento(Lancamento lancamento) {
+    public LancamentoResponseDTO atualizarLancamento(LancamentoRequestDTO lancamento) {
        
-        Lancamento lancamentoExistente = obterLancamentoPorId(lancamento.getId());
+        LancamentoResponseDTO lancamentoExistente = lancamentoRepository.findById(lancamento.getId())
+                .map(lancamentoMapper::paraLancamentoResponseDTO)
+                .orElse(null);
         if (lancamentoExistente == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Lançamento com o ID: " + lancamento.getId() + " não encontrado.");
         }
+
         lancamentoExistente.setValor(lancamento.getValor());
         lancamentoExistente.setObservacao(lancamento.getObservacao());
         lancamentoExistente.setEmpresa(lancamento.getEmpresa());
         lancamentoExistente.setBanco(lancamento.getBanco());
         lancamentoExistente.setDataLancamento(lancamento.getDataLancamento());
-        lancamentoExistente.setAtraso(lancamento.getAtraso());
         lancamentoExistente.setDataVencimento(lancamento.getDataVencimento());
-        return lancamentoRepository.save(lancamentoExistente);
+        lancamentoExistente.setAtraso(lancamento.getAtraso());
+        lancamentoExistente.setId(lancamento.getId());
+
+        LancamentoResponseDTO retorno = lancamentoMapper.paraLancamentoResponseDTO(lancamentoRepository.save(lancamentoMapper.paraLancamentoEntityRespDTO(lancamentoExistente)));    
+        if (retorno == null) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Erro ao atualizar o Lançamento. 🙁");
+        }
+        return retorno;
                    
     }
     
-    public List<Lancamento> obterLancamentosAtrasados() {
-        List<Lancamento> lancamentos = lancamentoRepository.findAll();
+    public List<LancamentoResponseDTO> obterLancamentosAtrasados() {
+        List<LancamentoResponseDTO> lancamentos = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.findAll());
         if (lancamentos.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado com atraso.");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado.");
         }
-        return lancamentos.stream()
+        List<LancamentoResponseDTO> retorno = lancamentos.stream()
                 .filter(lancamento -> lancamento.getAtraso())
                 .toList();
+        if (retorno.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado com atraso.");
+        }
+        return retorno;
     }
 
-    public List<Lancamento> obterLancamentosPorBanco(Long idBanco) {
-        List<Lancamento> lancamentos = lancamentoRepository.findAll();
+    public List<LancamentoResponseDTO> obterLancamentosPorBanco(Long idBanco) {
+        List<LancamentoResponseDTO> lancamentos = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.findAll());
         if (lancamentos.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado para o banco com ID: " + idBanco);
         }
@@ -79,29 +118,48 @@ public class LancamentoService {
                 .toList();  
     }
 
-    public List<Lancamento> obterLancamentosAtrasadosPorBanco(Long idBanco) {
-        List<Lancamento> lancamentos = lancamentoRepository.findAll();
+    public List<LancamentoResponseDTO> obterLancamentosAtrasadosPorBanco(Long idBanco) {
+        List<LancamentoResponseDTO> lancamentos = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.findAll());
         if (lancamentos.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado para o banco com ID: " + idBanco);
         }
-        return lancamentos.stream()
+        List<LancamentoResponseDTO> retorno = lancamentos.stream()
                 .filter(lancamento -> lancamento.getBanco().getId().equals(idBanco) && lancamento.getAtraso())
                 .toList();
+        if (retorno.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado para o banco com ID: " + idBanco + " com atraso.");
+        }
+        return retorno;
     }
 
     public void deletarLancamento(Long id) {
-        Lancamento lancamento = obterLancamentoPorId(id);
+        Lancamento lancamento = lancamentoRepository.findById(id).orElse(null);
+        if (lancamento == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Lançamento com o ID: " + id + " não encontrado.");
+        }   
         if (!lancamentoRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Erro ao deletar o Lançamento. 🙁");
         }
         lancamentoRepository.delete(lancamento);
     }
 
-    public List<Lancamento> obterLancamentosPorMes(int mes) {
-        List<Lancamento> lancamentos = lancamentoRepository.lancamentosNoMes(mes);
+    public List<LancamentoResponseDTO> obterLancamentosPorMes(int mes) {
+        List<LancamentoResponseDTO> lancamentos = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.lancamentosNoMes(mes));
         if (lancamentos.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado para o mês: " + mes);
         }
         return lancamentos;
     }
+
+    public BigDecimal obterTotalGastoNoMes(int mes) {
+        List<LancamentoResponseDTO> lancamentos = lancamentoMapper.paraLISTLancamentoResponseDTO(lancamentoRepository.lancamentosNoMes(mes));
+        if (lancamentos.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhum lançamento encontrado para o mês: " + mes);
+        }
+        BigDecimal totalMes = lancamentos.stream()
+            .map(LancamentoResponseDTO::getValor)
+            .reduce(BigDecimal.ZERO, BigDecimal::add);
+        return totalMes;
+    }
+
 }
